+++ Projects/Presentation/ViewModel/LoginViewModel.swift
@@ -92,6 +92,7 @@ class LoginViewModel {
         }
     }
     
+   
     func fetchUserProfile(accessToken: String, completion: @escaping (Result<UserResponse, Error>) -> Void) {
         self.useCase.fetchUserProfile(accessToken: accessToken, completion: completion)
     }
@@ -100,29 +101,3 @@ class LoginViewModel {
 
 
 
-protocol LoginRepository {
-    func fetchUserProfile(accessToken: String, completion: @escaping (Result<UserResponse, Error>) -> Void)
-}
-
-class LoginRepositoryImpl: LoginRepository {
-    func fetchUserProfile(accessToken: String, completion: @escaping (Result<UserResponse, any Error>) -> Void) {
-        let headers: HTTPHeaders = [
-            "Authorization" : "Bearer \(accessToken)"
-        ]
-        
-        AF.request(Endpoint.auth.urlString, method: .get, headers: headers)
-            .validate()
-            .responseDecodable(of: UserResponse.self) { response in
-                switch response.result {
-                case .success(let profile):
-                    // TODO: 여기서 유저 데이터 저장하기
-                    print(profile)
-                    
-                    completion(.success(profile))
-                    
-                case .failure(let error):
-                    completion(.failure(error))
-                }
-            }
-    }
-}
